initial commit. passes all test cases.
Used test cases to make some assumptions: The last maximum value in the score matrix (largest i, j) is used as the maximum score matrix value. In cases of equal scores, operations are prioritized in this order: match, mismatch, insertion in seq1, insertion in seq2, 0. Finally, ensured that match is not an option when seq1_i != seq2_j.

could decrease runtime and memory usage (delete gap matrix), but is still decently fast - ~1 min to align 100kbp sequence to 100bp sequence.